ref: make product abstract
